+++ assemblyadapter-common-recycler-divider/src/main/java/com/github/panpf/assemblyadapter/recycler/divider/IsFullSpanByPosition.kt
@@ -15,11 +15,9 @@
  */
 package com.github.panpf.assemblyadapter.recycler.divider
 
-import androidx.recyclerview.widget.RecyclerView
+import com.github.panpf.assemblyadapter.recycler.FullSpanSupport
 
 /**
  * Determining FullSpan based on position.
  */
-fun interface IsFullSpanByPosition {
-    fun isFullSpan(parent: RecyclerView, position: Int): Boolean
-}
+fun interface IsFullSpanByPosition : FullSpanSupport